Clean up Laravel defaults
One of the first things I do when I start a new Laravel project is clean up some of the things I know I won't be using, or will be replacing wholesale.

Laravel has done a fantastic job in 7+ in removing a lot of the cruft I would normally clean up, but there are still a few outliers I like to get out of the way so I can start from a clean slate.