Scaffold out project
When starting up a new project, I typically have a list of sane defaults I like to setup. This includes things like PHP Mess Detector, PHP CodeSniffer, a few tweaks for PHPUnit, and whatever I think I'll be needing for the frontend build process -- Tailwind and TailwindUI in this instance.