Implement User model and requirements.
I added a `players` local scope to the User model, since I know from the requirements we will need to filter out coaches.

In addition to the `PlayersIntegrityTest` specified in the requirements, I also added a `UserTest` to ensure the `players` scope worked correctly.

To assist in development and testing, I converted the provided SQL dump into JSON, added it the applications storage, and use that to seed the database.