Add User transformer
The meat of the requirements are contained in the `UserTransformer`.

This class accepts a collection of Users (players) and distributes them into an even number of teams of 18-22 players with similiar rankings, and assigns them a team name.

I used country codes as team names to add a bit of extra flair on the frontend.