Add frontend to to view teams
I used Tailwind CSS and TailwindUI to build a page for viewing the generated teams.

By keeping all of the complex business logic within the transformer, our controller is simply responsible for passing a collection of players to the transformer, and then passing the result on to the Blade page for display.